refactor(market): split oversized classes into focused support components
