+++ koduck-backend/src/main/java/com/koduck/controller/MarketAdvancedController.java
@@ -1,0 +1,420 @@
+package com.koduck.controller;
+
+import com.koduck.common.constants.ApiMessageConstants;
+import com.koduck.common.constants.ApiStatusCodeConstants;
+import com.koduck.common.constants.MarketConstants;
+import com.koduck.common.constants.PaginationConstants;
+import com.koduck.config.properties.DataServiceProperties;
+import com.koduck.dto.ApiResponse;
+import com.koduck.dto.market.BigOrderAlertDto;
+import com.koduck.dto.market.BigOrderStatsDto;
+import com.koduck.dto.market.CapitalRiverBreadthBandsDto;
+import com.koduck.dto.market.CapitalRiverBubbleDto;
+import com.koduck.dto.market.CapitalRiverDto;
+import com.koduck.dto.market.CapitalRiverTrackItemDto;
+import com.koduck.dto.market.CapitalRiverTracksDto;
+import com.koduck.dto.market.DataServiceResponse;
+import com.koduck.dto.market.PriceQuoteDto;
+import com.koduck.dto.market.SectorNetFlowDto;
+import com.koduck.dto.market.SectorNetFlowItemDto;
+import com.koduck.dto.market.SectorNetworkDto;
+import com.koduck.dto.market.SymbolInfoDto;
+import com.koduck.dto.market.TickDto;
+import com.koduck.service.MarketSectorNetFlowService;
+import com.koduck.service.MarketService;
+import com.koduck.service.SyntheticTickService;
+import com.koduck.service.TickStreamService;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotEmpty;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestClientException;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+import org.springframework.web.util.UriComponentsBuilder;
+
+/**
+ * Advanced market endpoints split from {@link MarketController} to keep controller classes focused.
+ *
+ * @author GitHub Copilot
+ * @date 2026-03-31
+ */
+@RestController
+@RequestMapping("/api/v1/market")
+@Tag(name = "市场数据", description = "市场高级数据接口")
+@Validated
+@Slf4j
+@RequiredArgsConstructor
+public class MarketAdvancedController {
+
+    private static final ZoneId MARKET_ZONE = ZoneId.of("Asia/Shanghai");
+    private static final String FEAR_GREED_INDEX_PATH =
+        System.getProperty("koduck.market.path.fearGreedIndex", "/market/fear-greed-index");
+    private static final String BREADTH_PATH =
+        System.getProperty("koduck.market.path.breadth", "/market/breadth");
+    private static final String BIG_ORDERS_PATH =
+        System.getProperty("koduck.market.path.bigOrders", "/market/big-orders");
+    private static final String BIG_ORDERS_STATS_PATH =
+        System.getProperty("koduck.market.path.bigOrdersStats", "/market/big-orders/stats");
+    private static final ParameterizedTypeReference<DataServiceResponse<Map<String, Object>>>
+        DATA_SERVICE_MAP_RESPONSE_TYPE = new ParameterizedTypeReference<>() {
+        };
+    private static final ParameterizedTypeReference<DataServiceResponse<List<BigOrderAlertDto>>>
+        BIG_ORDER_LIST_RESPONSE_TYPE = new ParameterizedTypeReference<>() {
+        };
+    private static final ParameterizedTypeReference<DataServiceResponse<BigOrderStatsDto>>
+        BIG_ORDER_STATS_RESPONSE_TYPE = new ParameterizedTypeReference<>() {
+        };
+
+    private final MarketService marketService;
+    private final MarketSectorNetFlowService marketSectorNetFlowService;
+    private final SyntheticTickService syntheticTickService;
+    private final TickStreamService tickStreamService;
+    private final DataServiceProperties dataServiceProperties;
+
+    @Qualifier("dataServiceRestTemplate")
+    private final RestTemplate dataServiceRestTemplate;
+
+    @GetMapping("/fear-greed-index")
+    public ApiResponse<Map<String, Object>> getFearGreedIndex() {
+        log.info("GET /api/v1/market/fear-greed-index");
+        if (!dataServiceProperties.isEnabled()) {
+            return ApiResponse.error(ApiStatusCodeConstants.SERVICE_UNAVAILABLE, ApiMessageConstants.DATA_SERVICE_DISABLED);
+        }
+        try {
+            ResponseEntity<DataServiceResponse<Map<String, Object>>> response = dataServiceRestTemplate.exchange(
+                dataServiceProperties.getBaseUrl() + FEAR_GREED_INDEX_PATH,
+                Objects.requireNonNull(HttpMethod.GET),
+                null,
+                Objects.requireNonNull(DATA_SERVICE_MAP_RESPONSE_TYPE));
+            DataServiceResponse<Map<String, Object>> body = response.getBody();
+            if (body == null || !body.isSuccess() || body.data() == null) {
+                return ApiResponse.error(ApiStatusCodeConstants.BAD_GATEWAY, ApiMessageConstants.FEAR_GREED_FETCH_FAILED);
+            }
+            return ApiResponse.success(body.data());
+        } catch (RestClientException e) {
+            log.warn("fear_greed_proxy_failed: {}", e.getMessage());
+            return ApiResponse.error(ApiStatusCodeConstants.BAD_GATEWAY, ApiMessageConstants.FEAR_GREED_FETCH_FAILED);
+        }
+    }
+
+    @GetMapping("/sector-net-flow")
+    public ApiResponse<SectorNetFlowDto> getSectorNetFlow(
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_MARKET) String market,
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_INDICATOR) String indicator,
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_LIST_LIMIT_STR) @Min(1) @Max(100) Integer limit,
+            @RequestParam(required = false)
+            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+            LocalDate tradeDate) {
+        log.info("GET /api/v1/market/sector-net-flow: market={}, indicator={}, limit={}, tradeDate={}",
+            market, indicator, limit, tradeDate);
+        SectorNetFlowDto result = tradeDate == null
+            ? marketSectorNetFlowService.getLatest(market, indicator, limit)
+            : marketSectorNetFlowService.getByTradeDate(market, indicator, tradeDate, limit);
+        if (result == null) {
+            return ApiResponse.error(ApiStatusCodeConstants.NOT_FOUND, ApiMessageConstants.SECTOR_NET_FLOW_NOT_FOUND);
+        }
+        return ApiResponse.success(result);
+    }
+
+    @GetMapping("/capital-river")
+    public ApiResponse<CapitalRiverDto> getCapitalRiver(
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_MARKET) String market,
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_INDICATOR) String indicator,
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_BUBBLE_COUNT_STR) @Min(1) @Max(10) Integer bubbleCount,
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_LIST_LIMIT_STR) @Min(1) @Max(100) Integer listLimit,
+            @RequestParam(required = false)
+            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+            LocalDate tradeDate) {
+        log.info("GET /api/v1/market/capital-river: market={}, indicator={}, bubbleCount={}, listLimit={}, tradeDate={}",
+            market, indicator, bubbleCount, listLimit, tradeDate);
+        SectorNetFlowDto snapshot = tradeDate == null
+            ? marketSectorNetFlowService.getLatest(market, indicator, listLimit)
+            : marketSectorNetFlowService.getByTradeDate(market, indicator, tradeDate, listLimit);
+        if (snapshot == null) {
+            return ApiResponse.error(ApiStatusCodeConstants.NOT_FOUND, ApiMessageConstants.CAPITAL_RIVER_NOT_FOUND);
+        }
+        List<CapitalRiverTrackItemDto> industry = toTrackItems(snapshot.industry());
+        List<CapitalRiverTrackItemDto> concept = toTrackItems(snapshot.concept());
+        List<CapitalRiverTrackItemDto> region = toTrackItems(snapshot.region());
+        List<CapitalRiverBubbleDto> bubbles = allTrackItems(industry, concept, region).stream()
+            .sorted(Comparator.comparing(this::absMainForceNet).reversed())
+            .limit(Math.max(1, bubbleCount))
+            .map(item -> new CapitalRiverBubbleDto(
+                item.sectorName(),
+                item.sectorType(),
+                item.mainForceNet(),
+                item.changePct()))
+            .toList();
+        CapitalRiverDto payload = new CapitalRiverDto(
+            snapshot.market(),
+            snapshot.indicator(),
+            snapshot.tradeDate(),
+            null,
+            null,
+            bubbles,
+            new CapitalRiverTracksDto(industry, concept, region),
+            new CapitalRiverBreadthBandsDto(
+                "Max Gainers (+10%)",
+                "Sector Breadth Distribution",
+                "Max Losers (-10%)",
+                List.of(100, 90, 82, 72, 60, 50, 40, 34, 28, 22, 16)
+            ),
+            snapshot.source(),
+            snapshot.quality());
+        return ApiResponse.success(payload);
+    }
+
+    @GetMapping("/breadth")
+    public ApiResponse<Map<String, Object>> getMarketBreadth() {
+        log.info("GET /api/v1/market/breadth");
+        if (!dataServiceProperties.isEnabled()) {
+            return ApiResponse.error(ApiStatusCodeConstants.SERVICE_UNAVAILABLE, ApiMessageConstants.DATA_SERVICE_DISABLED);
+        }
+        try {
+            ResponseEntity<DataServiceResponse<Map<String, Object>>> response = dataServiceRestTemplate.exchange(
+                dataServiceProperties.getBaseUrl() + BREADTH_PATH,
+                Objects.requireNonNull(HttpMethod.GET),
+                null,
+                Objects.requireNonNull(DATA_SERVICE_MAP_RESPONSE_TYPE));
+            DataServiceResponse<Map<String, Object>> body = response.getBody();
+            if (body == null || !body.isSuccess() || body.data() == null) {
+                return ApiResponse.error(ApiStatusCodeConstants.BAD_GATEWAY, ApiMessageConstants.BREADTH_FETCH_FAILED);
+            }
+            return ApiResponse.success(body.data());
+        } catch (RestClientException e) {
+            log.warn("breadth_proxy_failed: {}", e.getMessage());
+            return ApiResponse.error(ApiStatusCodeConstants.BAD_GATEWAY, ApiMessageConstants.BREADTH_FETCH_FAILED);
+        }
+    }
+
+    @GetMapping("/big-orders")
+    public ApiResponse<List<BigOrderAlertDto>> getBigOrders(
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_LIST_LIMIT_STR) @Min(1) @Max(50) Integer limit,
+            @RequestParam(name = "order_type", required = false) String orderType,
+            @RequestParam(name = "min_amount", defaultValue = "500000") @Min(0) Double minAmount) {
+        log.info("GET /api/v1/market/big-orders: limit={}, orderType={}, minAmount={}", limit, orderType, minAmount);
+        if (!dataServiceProperties.isEnabled()) {
+            log.warn("big_orders_proxy_skipped reason=data_service_disabled");
+            return ApiResponse.success(List.of());
+        }
+        try {
+            UriComponentsBuilder builder = UriComponentsBuilder
+                .fromUriString(dataServiceProperties.getBaseUrl() + BIG_ORDERS_PATH)
+                .queryParam("limit", limit)
+                .queryParam("min_amount", minAmount);
+            if (orderType != null && !orderType.isBlank()) {
+                builder.queryParam("order_type", orderType);
+            }
+            ResponseEntity<DataServiceResponse<List<BigOrderAlertDto>>> response = dataServiceRestTemplate.exchange(
+                builder.toUriString(),
+                Objects.requireNonNull(HttpMethod.GET),
+                null,
+                Objects.requireNonNull(BIG_ORDER_LIST_RESPONSE_TYPE));
+            DataServiceResponse<List<BigOrderAlertDto>> body = response.getBody();
+            if (body == null || !body.isSuccess() || body.data() == null) {
+                log.warn("big_orders_proxy_empty code={} message={}",
+                    body == null ? null : body.code(),
+                    body == null ? null : body.message());
+                return ApiResponse.success(List.of());
+            }
+            return ApiResponse.success(body.data());
+        } catch (RestClientException e) {
+            log.warn("big_orders_proxy_failed: {}", e.getMessage());
+            return ApiResponse.success(List.of());
+        }
+    }
+
+    @GetMapping("/big-orders/stats")
+    public ApiResponse<BigOrderStatsDto> getBigOrderStats() {
+        log.info("GET /api/v1/market/big-orders/stats");
+        if (!dataServiceProperties.isEnabled()) {
+            log.warn("big_order_stats_proxy_skipped reason=data_service_disabled");
+            return ApiResponse.success(BigOrderStatsDto.empty());
+        }
+        try {
+            ResponseEntity<DataServiceResponse<BigOrderStatsDto>> response = dataServiceRestTemplate.exchange(
+                dataServiceProperties.getBaseUrl() + BIG_ORDERS_STATS_PATH,
+                Objects.requireNonNull(HttpMethod.GET),
+                null,
+                Objects.requireNonNull(BIG_ORDER_STATS_RESPONSE_TYPE));
+            DataServiceResponse<BigOrderStatsDto> body = response.getBody();
+            if (body == null || !body.isSuccess() || body.data() == null) {
+                log.warn("big_order_stats_proxy_empty code={} message={}",
+                    body == null ? null : body.code(),
+                    body == null ? null : body.message());
+                return ApiResponse.success(BigOrderStatsDto.empty());
+            }
+            return ApiResponse.success(body.data());
+        } catch (RestClientException e) {
+            log.warn("big_order_stats_proxy_failed: {}", e.getMessage());
+            return ApiResponse.success(BigOrderStatsDto.empty());
+        }
+    }
+
+    @GetMapping("/hot")
+    public ApiResponse<List<SymbolInfoDto>> getHotStocks(
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_MARKET) String market,
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_PAGE_SIZE_STR)
+            @Min(value = 1, message = "每页数量最小为 1")
+            @Max(value = 100, message = "每页数量最大为 100")
+            Integer limit) {
+        log.info("GET /api/v1/market/hot?market={}&limit={}", market, limit);
+        List<SymbolInfoDto> hotStocks = marketService.getHotStocks(market, limit);
+        return ApiResponse.success(hotStocks);
+    }
+
+    @GetMapping("/sectors/network")
+    public ApiResponse<SectorNetworkDto> getSectorNetwork(
+            @RequestParam(defaultValue = MarketConstants.DEFAULT_MARKET) String market) {
+        log.info("GET /api/v1/market/sectors/network?market={}", market);
+        SectorNetworkDto network = marketService.getSectorNetwork(market);
+        return ApiResponse.success(network);
+    }
+
+    @GetMapping("/batch")
+    public ApiResponse<List<PriceQuoteDto>> getBatchPrices(
+            @RequestParam @NotEmpty(message = "股票代码列表不能为空")
+            @jakarta.validation.constraints.Size(max = 50, message = "股票代码最多 50 个")
+            List<String> symbols) {
+        log.info("GET /api/v1/market/batch: count={}", symbols.size());
+        List<PriceQuoteDto> quotes = marketService.getBatchPrices(symbols);
+        return ApiResponse.success(quotes);
+    }
+
+    @GetMapping("/ticks")
+    public ApiResponse<List<TickDto>> getTickData(
+            @RequestParam String market,
+            @RequestParam String symbol,
+            @RequestParam(defaultValue = PaginationConstants.DEFAULT_TICK_LIMIT_STR) @Min(1) @Max(500) Integer limit) {
+        log.info("GET /api/v1/market/ticks: market={}, symbol={}, limit={}", market, symbol, limit);
+        syntheticTickService.trackSymbol(symbol);
+        List<TickDto> historyTicks = syntheticTickService.getLatestTicks(symbol, limit);
+        return ApiResponse.success(historyTicks);
+    }
+
+    @GetMapping("/ticks/summary")
+    public ApiResponse<TickSummaryDto> getTickSummary(
+            @RequestParam String market,
+            @RequestParam String symbol) {
+        log.info("GET /api/v1/market/ticks/summary: market={}, symbol={}", market, symbol);
+        syntheticTickService.trackSymbol(symbol);
+        List<TickDto> historyTicks = syntheticTickService.getLatestTicks(symbol, 300);
+        if (historyTicks.isEmpty()) {
+            return ApiResponse.success(null);
+        }
+        long totalVolume = historyTicks.stream().mapToLong(TickDto::size).sum();
+        BigDecimal totalAmount = historyTicks.stream()
+            .mapToDouble(TickDto::amount)
+            .mapToObj(BigDecimal::valueOf)
+            .reduce(BigDecimal.ZERO, BigDecimal::add);
+        long buyVolume = historyTicks.stream()
+            .filter(tick -> "buy".equalsIgnoreCase(tick.type()))
+            .mapToLong(TickDto::size)
+            .sum();
+        long sellVolume = totalVolume - buyVolume;
+        int blockOrderCount = (int) historyTicks.stream()
+            .filter(tick -> "BLOCK_ORDER".equalsIgnoreCase(tick.flag()))
+            .count();
+        int totalTrades = historyTicks.size();
+        double avgTradeSize = totalTrades > 0
+            ? BigDecimal.valueOf(totalVolume)
+            .divide(BigDecimal.valueOf(totalTrades), 4, RoundingMode.HALF_UP)
+            .doubleValue()
+            : 0D;
+        TickDto latest = historyTicks.get(0);
+        TickSummaryDto summary = new TickSummaryDto(
+            symbol,
+            market,
+            totalTrades,
+            totalVolume,
+            totalAmount,
+            buyVolume,
+            sellVolume,
+            blockOrderCount,
+            avgTradeSize,
+            formatTickTimestampIso(latest.epochMillis()));
+        return ApiResponse.success(summary);
+    }
+
+    @GetMapping(value = "/ticks/stream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
+    public SseEmitter streamTicks(@RequestParam String market, @RequestParam String symbol) {
+        log.info("GET /api/v1/market/ticks/stream: market={}, symbol={}", market, symbol);
+        syntheticTickService.trackSymbol(symbol);
+        return tickStreamService.subscribe(symbol);
+    }
+
+    public record TickSummaryDto(
+        String symbol,
+        String market,
+        int totalTrades,
+        long totalVolume,
+        BigDecimal totalAmount,
+        long buyVolume,
+        long sellVolume,
+        int blockOrderCount,
+        double avgTradeSize,
+        String lastUpdated
+    ) {
+    }
+
+    private String formatTickTimestampIso(Long epochMillis) {
+        if (epochMillis == null) {
+            return LocalDateTime.now(MARKET_ZONE).toString();
+        }
+        return LocalDateTime.ofInstant(Instant.ofEpochMilli(epochMillis), MARKET_ZONE).toString();
+    }
+
+    private List<CapitalRiverTrackItemDto> toTrackItems(List<SectorNetFlowItemDto> items) {
+        if (items == null || items.isEmpty()) {
+            return List.of();
+        }
+        return items.stream()
+            .map(item -> new CapitalRiverTrackItemDto(
+                item.sectorName(),
+                item.sectorType(),
+                item.mainForceNet(),
+                item.changePct()))
+            .toList();
+    }
+
+    private List<CapitalRiverTrackItemDto> allTrackItems(
+            List<CapitalRiverTrackItemDto> industry,
+            List<CapitalRiverTrackItemDto> concept,
+            List<CapitalRiverTrackItemDto> region) {
+        List<CapitalRiverTrackItemDto> all = new ArrayList<>();
+        all.addAll(industry);
+        all.addAll(concept);
+        all.addAll(region);
+        return all;
+    }
+
+    private BigDecimal absMainForceNet(CapitalRiverTrackItemDto item) {
+        BigDecimal net = item.mainForceNet();
+        return net == null ? BigDecimal.ZERO : net.abs();
+    }
+}
