+++ koduck-backend/src/main/java/com/koduck/service/impl/MarketServiceImpl.java
@@ -13,13 +13,10 @@ import com.koduck.entity.StockBasic;
 import com.koduck.entity.StockRealtime;
 import com.koduck.repository.StockBasicRepository;
 import com.koduck.repository.StockRealtimeRepository;
-import com.koduck.market.MarketType;
-import com.koduck.market.provider.MarketDataProvider;
-import com.koduck.market.provider.ProviderFactory;
-import com.koduck.service.KlineService;
 import com.koduck.service.MarketService;
 import com.koduck.service.StockCacheService;
-import com.koduck.service.market.AKShareDataProvider;
+import com.koduck.service.support.MarketFallbackSupport;
+import com.koduck.service.support.MarketServiceSupport;
 import com.koduck.util.SymbolUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
@@ -27,16 +24,12 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.time.ZoneOffset;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -65,20 +58,20 @@ public class MarketServiceImpl implements MarketService {
     private final StockRealtimeRepository stockRealtimeRepository;
     private final StockBasicRepository stockBasicRepository;
     private final StockCacheService stockCacheService;
-    private final KlineService klineService;
-    private final ProviderFactory providerFactory;
+    private final MarketServiceSupport marketServiceSupport;
+    private final MarketFallbackSupport marketFallbackSupport;
     
     public MarketServiceImpl(
             StockRealtimeRepository stockRealtimeRepository,
             StockBasicRepository stockBasicRepository,
             StockCacheService stockCacheService,
-            KlineService klineService,
-            ProviderFactory providerFactory) {
+            MarketServiceSupport marketServiceSupport,
+            MarketFallbackSupport marketFallbackSupport) {
         this.stockRealtimeRepository = stockRealtimeRepository;
         this.stockBasicRepository = stockBasicRepository;
         this.stockCacheService = stockCacheService;
-        this.klineService = klineService;
-        this.providerFactory = providerFactory;
+        this.marketServiceSupport = marketServiceSupport;
+        this.marketFallbackSupport = marketFallbackSupport;
     }
     
     /**
@@ -121,10 +114,10 @@ public class MarketServiceImpl implements MarketService {
         // duplicate rows such as "002885" and "2885".
         Map<String, SymbolInfoDto> deduplicated = new LinkedHashMap<>();
         for (StockBasic basic : basics) {
-            SymbolInfoDto dto = mapToSymbolInfoDto(basic, realtimeMap.get(basic.getSymbol()));
+            SymbolInfoDto dto = marketServiceSupport.mapToSymbolInfoDto(basic, realtimeMap.get(basic.getSymbol()));
             String key = dto.market() + ":" + dto.symbol();
             SymbolInfoDto existing = deduplicated.get(key);
-            if (existing == null || shouldReplaceSymbol(existing, dto)) {
+            if (existing == null || marketServiceSupport.shouldReplaceSymbol(existing, dto)) {
                 deduplicated.put(key, dto);
             }
         }
@@ -133,39 +126,7 @@ public class MarketServiceImpl implements MarketService {
     }
 
     private List<SymbolInfoDto> searchSymbolsFromProvider(String keyword, int size) {
-        List<MarketDataProvider.SymbolInfo> providerResults = getAShareProvider()
-                .map(provider -> provider.searchSymbols(keyword, size))
-                .orElse(Collections.emptyList());
-        if (providerResults == null || providerResults.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        Map<String, SymbolInfoDto> deduplicated = new LinkedHashMap<>();
-        for (MarketDataProvider.SymbolInfo symbolInfo : providerResults) {
-            if (symbolInfo == null || symbolInfo.symbol() == null || symbolInfo.symbol().isBlank()) {
-                continue;
-            }
-            String normalizedSymbol = SymbolUtils.normalize(symbolInfo.symbol());
-            String market = symbolInfo.market() == null || symbolInfo.market().isBlank()
-                    ? DEFAULT_MARKET
-                    : symbolInfo.market();
-            String name = symbolInfo.name() == null || symbolInfo.name().isBlank()
-                    ? normalizedSymbol
-                    : symbolInfo.name();
-
-            SymbolInfoDto dto = SymbolInfoDto.builder()
-                    .symbol(normalizedSymbol)
-                    .name(name)
-                    .type(STOCK_TYPE)
-                    .market(market)
-                    .build();
-            deduplicated.putIfAbsent(market + ":" + normalizedSymbol, dto);
-        }
-
-        if (deduplicated.isEmpty()) {
-            return Collections.emptyList();
-        }
-        return deduplicated.values().stream().limit(size).toList();
+        return marketFallbackSupport.searchSymbolsFromProvider(keyword, size);
     }
     
     /**
@@ -188,13 +149,13 @@ public class MarketServiceImpl implements MarketService {
                     .findFirstBySymbolOrderByUpdatedAtDesc(symbol)
                     .orElse(null);
             if (entity == null) {
-                PriceQuoteDto fallbackQuote = tryBuildQuoteFromLatestKline(symbol);
+                PriceQuoteDto fallbackQuote = marketFallbackSupport.tryBuildQuoteFromLatestKline(symbol);
                 if (fallbackQuote != null) {
                     log.info("Built stock detail from latest kline data: symbol={}", symbol);
                     return fallbackQuote;
                 }
 
-                PriceQuoteDto providerQuote = fetchProviderPrice(symbol);
+                PriceQuoteDto providerQuote = marketFallbackSupport.fetchProviderPrice(symbol);
                 if (providerQuote != null) {
                     log.info("Fetched stock detail from data service: symbol={}", symbol);
                     return providerQuote;
@@ -207,16 +168,16 @@ public class MarketServiceImpl implements MarketService {
             log.debug("Found stock: symbol={}, name={}, price={}", 
                     entity.getSymbol(), entity.getName(), entity.getPrice());
             
-            return mapToPriceQuoteDto(entity);
+            return marketServiceSupport.mapToPriceQuoteDto(entity);
         } catch (Exception e) {
             log.error("Error getting stock detail: symbol={}, error={}", symbol, e.getMessage(), e);
-            PriceQuoteDto fallbackQuote = tryBuildQuoteFromLatestKline(symbol);
+            PriceQuoteDto fallbackQuote = marketFallbackSupport.tryBuildQuoteFromLatestKline(symbol);
             if (fallbackQuote != null) {
                 log.info("Recovered stock detail from latest kline after exception: symbol={}", symbol);
                 return fallbackQuote;
             }
 
-            PriceQuoteDto providerQuote = fetchProviderPrice(symbol);
+            PriceQuoteDto providerQuote = marketFallbackSupport.fetchProviderPrice(symbol);
             if (providerQuote != null) {
                 log.info("Recovered stock detail from data service after exception: symbol={}", symbol);
                 return providerQuote;
@@ -242,7 +203,7 @@ public class MarketServiceImpl implements MarketService {
         }
 
         try {
-            return fetchProviderValuation(symbol);
+            return marketFallbackSupport.fetchProviderValuation(symbol);
         } catch (Exception e) {
             log.error("Error getting stock valuation: symbol={}, error={}", symbol, e.getMessage(), e);
             throw e;
@@ -265,7 +226,7 @@ public class MarketServiceImpl implements MarketService {
         }
 
         try {
-            return fetchProviderIndustry(symbol);
+            return marketFallbackSupport.fetchProviderIndustry(symbol);
         } catch (Exception e) {
             log.error("Error getting stock industry: symbol={}, error={}", symbol, e.getMessage(), e);
             return null;
@@ -339,7 +300,7 @@ public class MarketServiceImpl implements MarketService {
         
         // Convert to DTOs
         List<PriceQuoteDto> dbQuotes = entities.stream()
-                .map(this::mapToPriceQuoteDto)
+            .map(marketServiceSupport::mapToPriceQuoteDto)
                 .toList();
         
         // Cache the database results
@@ -388,20 +349,20 @@ public class MarketServiceImpl implements MarketService {
                     .orElse(null);
             
             if (entity != null) {
-                return mapToStockStatsDto(entity, market);
+                return marketServiceSupport.mapToStockStatsDto(entity, market);
             }
             
             // Fallback: try to build from kline data
-            StockStatsDto klineStats = tryBuildStatsFromKline(symbol, market);
+            StockStatsDto klineStats = marketFallbackSupport.tryBuildStatsFromKline(symbol, market);
             if (klineStats != null) {
                 log.info("Built stock stats from kline data: symbol={}", symbol);
                 return klineStats;
             }
             
             // Final fallback: try data provider
-            PriceQuoteDto providerQuote = fetchProviderPrice(symbol);
+            PriceQuoteDto providerQuote = marketFallbackSupport.fetchProviderPrice(symbol);
             if (providerQuote != null) {
-                return mapPriceQuoteToStats(providerQuote, market);
+                return marketServiceSupport.mapPriceQuoteToStats(providerQuote, market);
             }
             
             log.warn("Stock stats not found for symbol={}", symbol);
@@ -411,14 +372,14 @@ public class MarketServiceImpl implements MarketService {
             log.error("Error getting stock stats: symbol={}, error={}", symbol, e.getMessage(), e);
             
             // Try fallback on exception
-            StockStatsDto klineStats = tryBuildStatsFromKline(symbol, market);
+            StockStatsDto klineStats = marketFallbackSupport.tryBuildStatsFromKline(symbol, market);
             if (klineStats != null) {
                 return klineStats;
             }
             
-            PriceQuoteDto providerQuote = fetchProviderPrice(symbol);
+            PriceQuoteDto providerQuote = marketFallbackSupport.fetchProviderPrice(symbol);
             if (providerQuote != null) {
-                return mapPriceQuoteToStats(providerQuote, market);
+                return marketServiceSupport.mapPriceQuoteToStats(providerQuote, market);
             }
             
             return null;
@@ -442,7 +403,7 @@ public class MarketServiceImpl implements MarketService {
         if (!indices.isEmpty()) {
             log.debug("Found {} indices in stock_realtime", indices.size());
             return indices.stream()
-                    .map(this::mapToMarketIndexDto)
+                    .map(marketServiceSupport::mapToMarketIndexDto)
                     .toList();
         }
         
@@ -459,7 +420,7 @@ public class MarketServiceImpl implements MarketService {
         log.debug("Found {} indices in stock_basic", basicIndices.size());
         // Map from StockBasic to MarketIndexDto (without price data)
         return basicIndices.stream()
-                .map(this::mapBasicToMarketIndexDto)
+            .map(marketServiceSupport::mapBasicToMarketIndexDto)
                 .toList();
     }
     
@@ -473,508 +434,15 @@ public class MarketServiceImpl implements MarketService {
      */
     @Override
     public List<SymbolInfoDto> getHotStocks(String market, int limit) {
-        log.debug("Getting hot stocks: market={}, limit={}", market, limit);
-        
-        try {
-            // Get stocks ordered by volume from database
-            List<StockRealtime> hotStocks = stockRealtimeRepository.findTopByVolume(limit);
-            
-            if (hotStocks.isEmpty()) {
-                log.warn("No hot stocks found in database");
-                return Collections.emptyList();
-            }
-            
-            // Get symbols for batch lookup of basic info
-            List<String> symbols = hotStocks.stream()
-                    .map(StockRealtime::getSymbol)
-                    .toList();
-            
-            // Get basic info for these stocks
-            List<StockBasic> basics = stockBasicRepository.findBySymbolIn(symbols);
-            Map<String, StockBasic> basicMap = basics.stream()
-                    .collect(Collectors.toMap(StockBasic::getSymbol, Function.identity(), (a, b) -> a));
-            
-            // Map to DTOs
-            return hotStocks.stream()
-                    .map(realtime -> mapRealtimeToSymbolInfoDto(realtime, basicMap.get(realtime.getSymbol()), market))
-                    .filter(Objects::nonNull)
-                    .limit(limit)
-                    .toList();
-                    
-        } catch (Exception e) {
-            log.error("Error getting hot stocks: market={}, limit={}, error={}", market, limit, e.getMessage(), e);
-            return Collections.emptyList();
-        }
+        return marketServiceSupport.getHotStocks(market, limit);
     }
     
-    private SymbolInfoDto mapRealtimeToSymbolInfoDto(StockRealtime realtime, StockBasic basic, String defaultMarket) {
-        if (realtime == null) {
-            return null;
-        }
-        
-        String market = defaultMarket;
-        String name = realtime.getName();
-        
-        if (basic != null) {
-            if (basic.getMarket() != null && !basic.getMarket().isBlank()) {
-                market = basic.getMarket();
-            }
-            if (basic.getName() != null && !basic.getName().isBlank()) {
-                name = basic.getName();
-            }
-        }
-        
-        return SymbolInfoDto.builder()
-                .symbol(realtime.getSymbol())
-                .name(name)
-                .type(realtime.getType())
-                .market(market)
-                .price(realtime.getPrice())
-                .changePercent(realtime.getChangePercent())
-                .volume(realtime.getVolume())
-                .amount(realtime.getAmount())
-                .build();
-    }
-
-    private Optional<MarketDataProvider> getAShareProvider() {
-        return providerFactory.getPrimaryProvider(MarketType.A_SHARE);
-    }
-
-    private PriceQuoteDto fetchProviderPrice(String symbol) {
-        return getAShareProvider()
-                .flatMap(this::toAkShareProvider)
-                .map(provider -> provider.getPrice(symbol))
-                .orElse(null);
-    }
-
-    private StockValuationDto fetchProviderValuation(String symbol) {
-        return getAShareProvider()
-                .flatMap(this::toAkShareProvider)
-                .map(provider -> provider.getStockValuation(symbol))
-                .orElse(null);
-    }
-
-    private StockIndustryDto fetchProviderIndustry(String symbol) {
-        return getAShareProvider()
-                .flatMap(this::toAkShareProvider)
-                .map(provider -> provider.getStockIndustry(symbol))
-                .orElse(null);
-    }
-
-    private Optional<AKShareDataProvider> toAkShareProvider(MarketDataProvider provider) {
-        if (provider instanceof AKShareDataProvider akShareProvider) {
-            return Optional.of(akShareProvider);
-        }
-        log.warn("Primary provider for A_SHARE is not AKShareDataProvider: {}", provider.getProviderName());
-        return Optional.empty();
-    }
-    
-    // ============ Mapping Methods ============
-    
-    private SymbolInfoDto mapToSymbolInfoDto(StockBasic basic, StockRealtime realtime) {
-        String normalizedSymbol = SymbolUtils.normalize(basic.getSymbol());
-
-        if (realtime != null) {
-            return SymbolInfoDto.builder()
-                    .symbol(normalizedSymbol)
-                    .name(basic.getName())
-                    .type(realtime.getType())
-                    .market(basic.getMarket())
-                    .price(realtime.getPrice())
-                    .changePercent(realtime.getChangePercent())
-                    .volume(realtime.getVolume())
-                    .amount(realtime.getAmount())
-                    .build();
-        }
-        
-        // Return basic info only if no realtime data
-        return SymbolInfoDto.builder()
-            .symbol(normalizedSymbol)
-                .name(basic.getName())
-                .type(STOCK_TYPE)
-                .market(basic.getMarket())
-                .build();
-    }
-
-    private boolean shouldReplaceSymbol(SymbolInfoDto existing, SymbolInfoDto candidate) {
-        if (existing.price() == null && candidate.price() != null) {
-            return true;
-        }
-        if (existing.changePercent() == null && candidate.changePercent() != null) {
-            return true;
-        }
-        if (existing.volume() == null && candidate.volume() != null) {
-            return true;
-        }
-        return existing.amount() == null && candidate.amount() != null;
-    }
-    
-    private PriceQuoteDto mapToPriceQuoteDto(StockRealtime entity) {
-        try {
-            return PriceQuoteDto.builder()
-                    .symbol(entity.getSymbol())
-                    .name(entity.getName())
-                    .type(entity.getType())
-                    .price(entity.getPrice())
-                    .open(entity.getOpenPrice())
-                    .high(entity.getHigh())
-                    .low(entity.getLow())
-                    .prevClose(entity.getPrevClose())
-                    .volume(entity.getVolume())
-                    .amount(entity.getAmount())
-                    .change(entity.getChangeAmount())
-                    .changePercent(entity.getChangePercent())
-                    .bidPrice(entity.getBidPrice())
-                    .bidVolume(entity.getBidVolume())
-                    .askPrice(entity.getAskPrice())
-                    .askVolume(entity.getAskVolume())
-                    .timestamp(entity.getUpdatedAt() != null ? entity.getUpdatedAt().toInstant(ZoneOffset.UTC) : null)
-                    .build();
-        } catch (Exception e) {
-            log.error("Failed to map StockRealtime to PriceQuoteDto: symbol={}, error={}", 
-                    entity != null ? entity.getSymbol() : "null", e.getMessage(), e);
-            throw e;
-        }
-    }
-
-    private PriceQuoteDto buildQuoteFromLatestKline(String symbol) {
-        StockBasic basic = stockBasicRepository.findBySymbol(symbol).orElse(null);
-        if (basic == null) {
-            return null;
-        }
-
-        String market = basic.getMarket() != null && !basic.getMarket().isBlank()
-                ? basic.getMarket()
-                : DEFAULT_MARKET;
-
-        PriceQuoteDto quote = buildQuoteFromKline(symbol, basic, market);
-        if (quote != null) {
-            return quote;
-        }
-
-        if (!DEFAULT_MARKET.equals(market)) {
-            return buildQuoteFromKline(symbol, basic, DEFAULT_MARKET);
-        }
-
-        return null;
-    }
-
-    private PriceQuoteDto tryBuildQuoteFromLatestKline(String symbol) {
-        try {
-            return buildQuoteFromLatestKline(symbol);
-        } catch (Exception ex) {
-            log.warn("Failed to build quote from kline: symbol={}, error={}", symbol, ex.getMessage());
-            return null;
-        }
-    }
-
-    private PriceQuoteDto buildQuoteFromKline(String symbol, StockBasic basic, String market) {
-        List<com.koduck.dto.market.KlineDataDto> recent =
-            klineService.getKlineData(market, symbol, DAILY_TIMEFRAME, 2, null);
-        recent = normalizeKlineData(recent);
-        if (recent.isEmpty()) {
-            return null;
-        }
-
-        com.koduck.dto.market.KlineDataDto latest = recent.get(recent.size() - 1);
-        BigDecimal prevClose = recent.size() >= 2 ? recent.get(recent.size() - 2).close() : null;
-        BigDecimal price = latest.close();
-        BigDecimal change = calculateChange(price, prevClose);
-        BigDecimal changePercent = calculateChangePercent(change, prevClose);
-
-        return PriceQuoteDto.builder()
-            .symbol(symbol)
-            .name(basic.getName())
-            .type(STOCK_TYPE)
-            .price(price)
-            .open(latest.open())
-            .high(latest.high())
-            .low(latest.low())
-            .prevClose(prevClose)
-            .volume(latest.volume())
-            .amount(latest.amount())
-            .change(change)
-            .changePercent(changePercent)
-            .timestamp(latest.timestamp() != null ? Instant.ofEpochSecond(latest.timestamp()) : null)
-            .build();
-    }
-
-    private List<com.koduck.dto.market.KlineDataDto> normalizeKlineData(
-            List<com.koduck.dto.market.KlineDataDto> rawData) {
-        if (rawData == null || rawData.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        List<com.koduck.dto.market.KlineDataDto> normalized = new ArrayList<>();
-        for (Object item : rawData) {
-            if (item instanceof com.koduck.dto.market.KlineDataDto dto) {
-                normalized.add(dto);
-                continue;
-            }
-            if (item instanceof Map<?, ?> map) {
-                normalized.add(
-                    com.koduck.dto.market.KlineDataDto.builder()
-                        .timestamp(toLong(map.get("timestamp")))
-                        .open(toBigDecimal(map.get("open")))
-                        .high(toBigDecimal(map.get("high")))
-                        .low(toBigDecimal(map.get("low")))
-                        .close(toBigDecimal(map.get("close")))
-                        .volume(toLong(map.get("volume")))
-                        .amount(toBigDecimal(map.get("amount")))
-                        .build()
-                );
-            }
-        }
-        return normalized;
-    }
-
-    private BigDecimal toBigDecimal(Object value) {
-        if (value == null) {
-            return null;
-        }
-        if (value instanceof BigDecimal bigDecimal) {
-            return bigDecimal;
-        }
-        if (value instanceof Number number) {
-            return BigDecimal.valueOf(number.doubleValue());
-        }
-        try {
-            return new BigDecimal(value.toString());
-        } catch (NumberFormatException _) {
-            return null;
-        }
-    }
-
-    private Long toLong(Object value) {
-        if (value == null) {
-            return null;
-        }
-        if (value instanceof Number number) {
-            return number.longValue();
-        }
-        try {
-            return Long.parseLong(value.toString());
-        } catch (NumberFormatException _) {
-            return null;
-        }
-    }
-
-    private BigDecimal calculateChange(BigDecimal price, BigDecimal prevClose) {
-        if (price == null || prevClose == null) {
-            return null;
-        }
-        return price.subtract(prevClose);
-    }
-
-    private BigDecimal calculateChangePercent(BigDecimal change, BigDecimal prevClose) {
-        if (change == null || prevClose == null || BigDecimal.ZERO.compareTo(prevClose) == 0) {
-            return null;
-        }
-        return change.multiply(BigDecimal.valueOf(100))
-                .divide(prevClose, 4, RoundingMode.HALF_UP);
-    }
-    
-    private MarketIndexDto mapToMarketIndexDto(StockRealtime entity) {
-        return MarketIndexDto.builder()
-                .symbol(entity.getSymbol())
-                .name(entity.getName())
-                .type(entity.getType())
-                .price(entity.getPrice())
-                .change(entity.getChangeAmount())
-                .changePercent(entity.getChangePercent())
-                .open(entity.getOpenPrice())
-                .high(entity.getHigh())
-                .low(entity.getLow())
-                .prevClose(entity.getPrevClose())
-                .volume(entity.getVolume())
-                .amount(entity.getAmount())
-                .timestamp(entity.getUpdatedAt() != null ? entity.getUpdatedAt().toInstant(ZoneOffset.ofHours(8)) : null)
-                .build();
-    }
-    
-    private MarketIndexDto mapBasicToMarketIndexDto(StockBasic basic) {
-        return MarketIndexDto.builder()
-                .symbol(basic.getSymbol())
-                .name(basic.getName())
-                .type(basic.getType())
-                .build();
-    }
-    
-    // ============ StockStats Helper Methods ============
-    
-    private StockStatsDto mapToStockStatsDto(StockRealtime entity, String market) {
-        return StockStatsDto.builder()
-                .symbol(entity.getSymbol())
-                .market(market)
-                .open(entity.getOpenPrice())
-                .high(entity.getHigh())
-                .low(entity.getLow())
-                .current(entity.getPrice())
-                .prevClose(entity.getPrevClose())
-                .change(entity.getChangeAmount())
-                .changePercent(entity.getChangePercent())
-                .volume(entity.getVolume())
-                .amount(entity.getAmount())
-                .timestamp(entity.getUpdatedAt() != null ? entity.getUpdatedAt().toInstant(ZoneOffset.UTC) : null)
-                .build();
-    }
-    
-    private StockStatsDto mapPriceQuoteToStats(PriceQuoteDto quote, String market) {
-        return StockStatsDto.builder()
-                .symbol(quote.symbol())
-                .market(market)
-                .open(quote.open())
-                .high(quote.high())
-                .low(quote.low())
-                .current(quote.price())
-                .prevClose(quote.prevClose())
-                .change(quote.change())
-                .changePercent(quote.changePercent())
-                .volume(quote.volume())
-                .amount(quote.amount())
-                .timestamp(quote.timestamp())
-                .build();
-    }
-    
-    private StockStatsDto tryBuildStatsFromKline(String symbol, String market) {
-        try {
-            StockBasic basic = stockBasicRepository.findBySymbol(symbol).orElse(null);
-            if (basic == null) {
-                return null;
-            }
-
-            String actualMarket = market;
-            if (actualMarket == null || actualMarket.isBlank()) {
-                String basicMarket = basic.getMarket();
-                if (basicMarket != null && !basicMarket.isBlank()) {
-                    actualMarket = basicMarket;
-                } else {
-                    actualMarket = DEFAULT_MARKET;
-                }
-            }
-            
-            List<com.koduck.dto.market.KlineDataDto> recent =
-                    klineService.getKlineData(actualMarket, symbol, DAILY_TIMEFRAME, 2, null);
-            recent = normalizeKlineData(recent);
-            
-            if (recent.isEmpty()) {
-                return null;
-            }
-            
-            com.koduck.dto.market.KlineDataDto latest = recent.get(recent.size() - 1);
-            BigDecimal prevClose = recent.size() >= 2 ? recent.get(recent.size() - 2).close() : null;
-            BigDecimal current = latest.close();
-            BigDecimal change = calculateChange(current, prevClose);
-            BigDecimal changePercent = calculateChangePercent(change, prevClose);
-            
-            return StockStatsDto.builder()
-                    .symbol(symbol)
-                    .market(actualMarket)
-                    .open(latest.open())
-                    .high(latest.high())
-                    .low(latest.low())
-                    .current(current)
-                    .prevClose(prevClose)
-                    .change(change)
-                    .changePercent(changePercent)
-                    .volume(latest.volume())
-                    .amount(latest.amount())
-                    .timestamp(latest.timestamp() != null ? Instant.ofEpochSecond(latest.timestamp()) : null)
-                    .build();
-        } catch (Exception ex) {
-            log.warn("Failed to build stats from kline: symbol={}, error={}", symbol, ex.getMessage());
-            return null;
-        }
-    }
-
     // ============ Sector Network Methods ============
 
     @Override
     public SectorNetworkDto getSectorNetwork(String market) {
         log.debug("Getting sector network data for market: {}", market);
-        
-        // For now, return mock data
-        // In production, this should fetch from industry/sector data tables
-        // and calculate correlations based on historical price data
-        return generateMockSectorNetwork();
+        return marketServiceSupport.generateMockSectorNetwork(POSITIVE_LINK_TYPE, NEGATIVE_LINK_TYPE);
     }
     
-    private SectorNetworkDto generateMockSectorNetwork() {
-        var nodes = List.of(
-            SectorNetworkDto.SectorNode.builder()
-                    .id("1").name("新能源").marketCap(new BigDecimal("8500"))
-                    .flow(new BigDecimal("67.3")).change(new BigDecimal("3.2")).group(1).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("2").name("锂电池").marketCap(new BigDecimal("4200"))
-                    .flow(new BigDecimal("34.2")).change(new BigDecimal("2.8")).group(1).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("3").name("光伏").marketCap(new BigDecimal("3800"))
-                    .flow(new BigDecimal("28.5")).change(new BigDecimal("2.1")).group(1).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("4").name("储能").marketCap(new BigDecimal("2900"))
-                    .flow(new BigDecimal("15.8")).change(new BigDecimal("1.9")).group(1).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("5").name("银行").marketCap(new BigDecimal("12000"))
-                    .flow(new BigDecimal("45.2")).change(new BigDecimal("1.2")).group(2).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("6").name("保险").marketCap(new BigDecimal("5600"))
-                    .flow(new BigDecimal("12.3")).change(new BigDecimal("0.8")).group(2).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("7").name("证券").marketCap(new BigDecimal("4800"))
-                    .flow(new BigDecimal("-8.5")).change(new BigDecimal("-0.5")).group(2).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("8").name("科技").marketCap(new BigDecimal("9200"))
-                    .flow(new BigDecimal("-67.5")).change(new BigDecimal("-2.8")).group(3).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("9").name("半导体").marketCap(new BigDecimal("6500"))
-                    .flow(new BigDecimal("-45.2")).change(new BigDecimal("-2.1")).group(3).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("10").name("软件").marketCap(new BigDecimal("3800"))
-                    .flow(new BigDecimal("-22.3")).change(new BigDecimal("-1.5")).group(3).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("11").name("医药").marketCap(new BigDecimal("7200"))
-                    .flow(new BigDecimal("-12.1")).change(new BigDecimal("-0.8")).group(4).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("12").name("医疗器械").marketCap(new BigDecimal("3400"))
-                    .flow(new BigDecimal("8.5")).change(new BigDecimal("0.5")).group(4).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("13").name("消费").marketCap(new BigDecimal("6800"))
-                    .flow(new BigDecimal("23.4")).change(new BigDecimal("1.5")).group(5).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("14").name("白酒").marketCap(new BigDecimal("5200"))
-                    .flow(new BigDecimal("18.9")).change(new BigDecimal("1.2")).group(5).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("15").name("汽车").marketCap(new BigDecimal("5800"))
-                    .flow(new BigDecimal("15.6")).change(new BigDecimal("0.9")).group(6).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("16").name("军工").marketCap(new BigDecimal("4200"))
-                    .flow(new BigDecimal("12.8")).change(new BigDecimal("0.7")).group(7).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("17").name("地产").marketCap(new BigDecimal("3600"))
-                    .flow(new BigDecimal("-45.6")).change(new BigDecimal("-3.2")).group(8).build(),
-            SectorNetworkDto.SectorNode.builder()
-                    .id("18").name("建材").marketCap(new BigDecimal("2800"))
-                    .flow(new BigDecimal("-18.9")).change(new BigDecimal("-1.8")).group(8).build()
-        );
-        
-        var links = List.of(
-            SectorNetworkDto.SectorLink.builder().source("1").target("2").value(new BigDecimal("0.85")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("1").target("3").value(new BigDecimal("0.78")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("1").target("4").value(new BigDecimal("0.72")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("2").target("3").value(new BigDecimal("0.65")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("5").target("6").value(new BigDecimal("0.68")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("5").target("7").value(new BigDecimal("0.55")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("8").target("9").value(new BigDecimal("0.82")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("8").target("10").value(new BigDecimal("0.75")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("9").target("10").value(new BigDecimal("0.70")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("11").target("12").value(new BigDecimal("0.62")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("13").target("14").value(new BigDecimal("0.58")).type(POSITIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("1").target("8").value(new BigDecimal("-0.65")).type(NEGATIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("1").target("5").value(new BigDecimal("-0.45")).type(NEGATIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("5").target("8").value(new BigDecimal("-0.55")).type(NEGATIVE_LINK_TYPE).build(),
-            SectorNetworkDto.SectorLink.builder().source("2").target("9").value(new BigDecimal("-0.48")).type(NEGATIVE_LINK_TYPE).build()
-        );
-        
-        return SectorNetworkDto.builder().nodes(nodes).links(links).build();
-    }
 }
