+++ koduck-backend/src/main/java/com/koduck/service/market/AKShareDataProvider.java
@@ -11,7 +11,7 @@ import com.koduck.market.MarketType;
 import com.koduck.market.model.KlineData;
 import com.koduck.market.model.TickData;
 import com.koduck.market.provider.MarketDataProvider;
-import com.koduck.market.util.DataConverter;
+import com.koduck.service.market.support.AKShareDataMapperSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -45,13 +45,7 @@ public class AKShareDataProvider implements MarketDataProvider {
     private static final String DATA_SERVICE_DISABLED_MESSAGE = "Data service is disabled";
     private static final String A_SHARE_BASE_PATH = "/a-share";
     private static final String KEY_SYMBOL = "symbol";
-    private static final String KEY_NAME = "name";
     private static final String KEY_LIMIT = "limit";
-    private static final String KEY_VOLUME = "volume";
-    private static final String KEY_AMOUNT = "amount";
-    private static final String EXCHANGE_SH = "SH";
-    private static final String EXCHANGE_SZ = "SZ";
-    private static final String EXCHANGE_CN = "CN";
     private static final String RESPONSE_TYPE_MESSAGE = "responseType must not be null";
     private static final String HTTP_GET_MESSAGE = "HTTP GET must not be null";
     private static final String HTTP_POST_MESSAGE = "HTTP POST must not be null";
@@ -135,7 +129,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getListDataResponseType()
                     );
             
-            return parseKlineResponse(response.getBody());
+            DataServiceResponse<List<Map<String, Object>>> body = response.getBody();
+            return AKShareDataMapperSupport.parseKlineResponse(body == null ? null : body.data());
             
         } catch (RestClientException e) {
             throw new MarketDataException("Failed to get kline data", e);
@@ -246,7 +241,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getListDataResponseType()
                     );
             
-            return parseSymbolInfoResponse(response.getBody());
+            DataServiceResponse<List<Map<String, Object>>> body = response.getBody();
+            return AKShareDataMapperSupport.parseSymbolInfoResponse(body == null ? null : body.data());
             
         } catch (RestClientException e) {
             log.error("Failed to search symbols: {}", e.getMessage());
@@ -278,7 +274,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getMapDataResponseType()
                     );
             
-            return parsePriceQuoteResponse(response.getBody());
+            DataServiceResponse<Map<String, Object>> body = response.getBody();
+            return AKShareDataMapperSupport.parsePriceQuoteResponse(body == null ? null : body.data());
             
         } catch (RestClientException e) {
             throw new ExternalServiceException("DataService",
@@ -318,7 +315,7 @@ public class AKShareDataProvider implements MarketDataProvider {
             }
             
             return body.data().stream()
-                    .map(this::mapToPriceQuoteDto)
+                    .map(AKShareDataMapperSupport::mapToPriceQuoteDto)
                     .toList();
             
         } catch (RestClientException e) {
@@ -349,7 +346,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getListDataResponseType()
                     );
             
-            return parseSymbolListResponse(response.getBody());
+            DataServiceResponse<List<Map<String, Object>>> body = response.getBody();
+            return AKShareDataMapperSupport.parseSymbolListResponse(body == null ? null : body.data());
             
         } catch (RestClientException e) {
             log.error("Failed to get hot symbols: {}", e.getMessage());
@@ -379,7 +377,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getMapDataResponseType()
                     );
 
-            return parseStockValuationResponse(response.getBody());
+            DataServiceResponse<Map<String, Object>> body = response.getBody();
+            return AKShareDataMapperSupport.parseStockValuationResponse(body == null ? null : body.data());
 
         } catch (RestClientException e) {
             throw new ExternalServiceException("DataService",
@@ -409,7 +408,8 @@ public class AKShareDataProvider implements MarketDataProvider {
                             getMapDataResponseType()
                     );
 
-            return parseStockIndustryResponse(response.getBody());
+            DataServiceResponse<Map<String, Object>> body = response.getBody();
+            return AKShareDataMapperSupport.parseStockIndustryResponse(body == null ? null : body.data());
 
         } catch (RestClientException e) {
             throw new ExternalServiceException("DataService",
@@ -417,200 +417,6 @@ public class AKShareDataProvider implements MarketDataProvider {
         }
     }
     
-    // Helper methods
-    
-    private List<KlineData> parseKlineResponse(DataServiceResponse<List<Map<String, Object>>> response) {
-        if (response == null || response.data() == null) {
-            return Collections.emptyList();
-        }
-        
-        return response.data().stream()
-                .map(this::mapToKlineData)
-                .toList();
-    }
-    
-    private KlineData mapToKlineData(Map<String, Object> data) {
-        return KlineData.builder()
-            .symbol(getString(data, KEY_SYMBOL))
-            .market(MarketType.A_SHARE.getCode())
-            .timestamp(DataConverter.toInstantFromMillis(getLong(data, "timestamp")))
-            .open(DataConverter.toBigDecimal(getString(data, "open")))
-            .high(DataConverter.toBigDecimal(getString(data, "high")))
-            .low(DataConverter.toBigDecimal(getString(data, "low")))
-            .close(DataConverter.toBigDecimal(getString(data, "close")))
-            .volume(getLong(data, KEY_VOLUME))
-            .amount(DataConverter.toBigDecimal(getString(data, KEY_AMOUNT)))
-            .timeframe(getString(data, "timeframe"))
-            .build();
-    }
-    
-    private List<SymbolInfo> parseSymbolInfoResponse(DataServiceResponse<List<Map<String, Object>>> response) {
-        if (response == null || response.data() == null) {
-            return Collections.emptyList();
-        }
-        
-        return response.data().stream()
-                .map(this::mapToSymbolInfo)
-                .toList();
-    }
-    
-    private SymbolInfo mapToSymbolInfo(Map<String, Object> data) {
-        String symbol = getString(data, KEY_SYMBOL);
-        // Normalize symbol with exchange suffix
-        String normalizedSymbol = DataConverter.normalizeSymbol(symbol, MarketType.A_SHARE.getCode());
-        String exchange = resolveExchange(normalizedSymbol);
-        
-        return new SymbolInfo(
-            normalizedSymbol,
-            getString(data, KEY_NAME),
-            MarketType.A_SHARE.getCode(),
-            exchange,
-            "stock"
-        );
-    }
-    
-    private List<SymbolInfoDto> parseSymbolListResponse(DataServiceResponse<List<Map<String, Object>>> response) {
-        if (response == null || response.data() == null) {
-            return Collections.emptyList();
-        }
-        
-        return response.data().stream()
-                .map(this::mapToSymbolInfoDto)
-                .toList();
-    }
-    
-    private SymbolInfoDto mapToSymbolInfoDto(Map<String, Object> data) {
-        return SymbolInfoDto.builder()
-                .symbol(getString(data, KEY_SYMBOL))
-                .name(getString(data, KEY_NAME))
-                .type(Optional.ofNullable(getString(data, "type")).orElse("STOCK"))
-                .market(getString(data, "market"))
-                .price(DataConverter.toBigDecimal(getString(data, "price")))
-                .changePercent(DataConverter.toBigDecimal(getString(data, "change_percent")))
-                .volume(getLong(data, KEY_VOLUME))
-                .amount(DataConverter.toBigDecimal(getString(data, KEY_AMOUNT)))
-                .build();
-    }
-    
-    private PriceQuoteDto parsePriceQuoteResponse(DataServiceResponse<Map<String, Object>> response) {
-        if (response == null || response.data() == null) {
-            return null;
-        }
-        
-        return mapToPriceQuoteDto(response.data());
-    }
-
-    private StockValuationDto parseStockValuationResponse(DataServiceResponse<Map<String, Object>> response) {
-        if (response == null || response.data() == null) {
-            return null;
-        }
-
-        return mapToStockValuationDto(response.data());
-    }
-
-    private StockIndustryDto parseStockIndustryResponse(DataServiceResponse<Map<String, Object>> response) {
-        if (response == null || response.data() == null) {
-            return null;
-        }
-
-        return mapToStockIndustryDto(response.data());
-    }
-    
-    private PriceQuoteDto mapToPriceQuoteDto(Map<String, Object> data) {
-        return PriceQuoteDto.builder()
-                .symbol(getString(data, KEY_SYMBOL))
-                .name(getString(data, KEY_NAME))
-                .type(Optional.ofNullable(getString(data, "type")).orElse("STOCK"))
-                .price(DataConverter.toBigDecimal(getString(data, "price")))
-                .open(DataConverter.toBigDecimal(getString(data, "open")))
-                .high(DataConverter.toBigDecimal(getString(data, "high")))
-                .low(DataConverter.toBigDecimal(getString(data, "low")))
-                .prevClose(DataConverter.toBigDecimal(getString(data, "prev_close")))
-                .volume(getLong(data, KEY_VOLUME))
-                .amount(DataConverter.toBigDecimal(getString(data, KEY_AMOUNT)))
-                .change(DataConverter.toBigDecimal(getString(data, "change")))
-                .changePercent(DataConverter.toBigDecimal(getString(data, "change_percent")))
-                .bidPrice(DataConverter.toBigDecimal(getString(data, "bid_price")))
-                .bidVolume(getLong(data, "bid_volume"))
-                .askPrice(DataConverter.toBigDecimal(getString(data, "ask_price")))
-                .askVolume(getLong(data, "ask_volume"))
-                .timestamp(getInstant(data, "timestamp"))
-                .build();
-    }
-
-    private StockValuationDto mapToStockValuationDto(Map<String, Object> data) {
-        return StockValuationDto.builder()
-                .symbol(getString(data, KEY_SYMBOL))
-                .name(getString(data, KEY_NAME))
-                .peTtm(DataConverter.toBigDecimal(getString(data, "pe_ttm")))
-                .pb(DataConverter.toBigDecimal(getString(data, "pb")))
-                .psTtm(DataConverter.toBigDecimal(getString(data, "ps_ttm")))
-                .marketCap(DataConverter.toBigDecimal(getString(data, "market_cap")))
-                .floatMarketCap(DataConverter.toBigDecimal(getString(data, "float_market_cap")))
-                .totalShares(DataConverter.toBigDecimal(getString(data, "total_shares")))
-                .floatShares(DataConverter.toBigDecimal(getString(data, "float_shares")))
-                .floatRatio(DataConverter.toBigDecimal(getString(data, "float_ratio")))
-                .turnoverRate(DataConverter.toBigDecimal(getString(data, "turnover_rate")))
-                .build();
-    }
-
-    private StockIndustryDto mapToStockIndustryDto(Map<String, Object> data) {
-        return StockIndustryDto.builder()
-                .symbol(getString(data, KEY_SYMBOL))
-                .name(getString(data, KEY_NAME))
-                .industry(getString(data, "industry"))
-                .sector(getString(data, "sector"))
-                .subIndustry(getString(data, "sub_industry"))
-                .board(getString(data, "board"))
-                .build();
-    }
-    
-    // Helper methods for safe type conversion
-    private String getString(Map<String, Object> data, String key) {
-        Object value = data.get(key);
-        return value != null ? value.toString() : null;
-    }
-    
-    private Long getLong(Map<String, Object> data, String key) {
-        Object value = data.get(key);
-        if (value == null) {
-            return null;
-        }
-        if (value instanceof Number number) {
-            return number.longValue();
-        }
-        try {
-            return Long.parseLong(value.toString());
-        } catch (NumberFormatException _) {
-            return null;
-        }
-    }
-    
-    private Instant getInstant(Map<String, Object> data, String key) {
-        Object value = data.get(key);
-        if (value == null) {
-            return null;
-        }
-        if (value instanceof String timestamp) {
-            try {
-                return Instant.parse(timestamp);
-            } catch (Exception _) {
-                return null;
-            }
-        }
-        return null;
-    }
-
-    private String resolveExchange(String normalizedSymbol) {
-        if (normalizedSymbol.contains(".SH")) {
-            return EXCHANGE_SH;
-        }
-        if (normalizedSymbol.contains(".SZ")) {
-            return EXCHANGE_SZ;
-        }
-        return EXCHANGE_CN;
-    }
-
     private static @NonNull HttpMethod getHttpGet() {
         return Objects.requireNonNull(HttpMethod.GET, HTTP_GET_MESSAGE);
     }
