+++ koduck-backend/src/main/java/com/koduck/service/market/USStockProvider.java
@@ -5,6 +5,7 @@ import com.koduck.market.MarketType;
 import com.koduck.market.model.KlineData;
 import com.koduck.market.model.TickData;
 import com.koduck.market.provider.MarketDataProvider;
+import com.koduck.service.market.support.USStockMockDataProvider;
 import lombok.Getter;
 import lombok.Setter;
 import org.slf4j.Logger;
@@ -19,11 +20,9 @@ import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.time.*;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * US Stock market data provider using Finnhub API.
@@ -52,7 +51,7 @@ public class USStockProvider implements MarketDataProvider {
     private final Set<String> subscribedSymbols = ConcurrentHashMap.newKeySet();
     
     // Fallback to mock data when API is not available
-    private final MockDataProvider mockProvider = new MockDataProvider();
+    private final USStockMockDataProvider mockProvider = new USStockMockDataProvider();
 
     public USStockProvider(
             FinnhubProperties properties,
@@ -392,156 +391,4 @@ public class USStockProvider implements MarketDataProvider {
     record SearchResult(String description, String displaySymbol, String symbol, String type, String exchange) {
     }
     
-    /**
-     * Mock data provider for fallback when API is not available
-     */
-    private static class MockDataProvider {
-        private final Map<String, BigDecimal> basePrices = new HashMap<>();
-        
-        MockDataProvider() {
-            basePrices.put("AAPL", new BigDecimal("175.50"));
-            basePrices.put("MSFT", new BigDecimal("420.00"));
-            basePrices.put("GOOGL", new BigDecimal("165.00"));
-            basePrices.put("AMZN", new BigDecimal("180.00"));
-            basePrices.put("TSLA", new BigDecimal("240.00"));
-            basePrices.put("META", new BigDecimal("500.00"));
-            basePrices.put("NVDA", new BigDecimal("880.00"));
-            basePrices.put("AMD", new BigDecimal("180.00"));
-            basePrices.put("INTC", new BigDecimal("45.00"));
-            basePrices.put("NFLX", new BigDecimal("600.00"));
-        }
-        
-        boolean isAvailable() {
-            return true;
-        }
-        
-        int getHealthScore() {
-            return 50;
-        }
-        
-        List<KlineData> getKlineData(String symbol, String timeframe, int limit,
-                                      Instant startTime, Instant endTime) {
-            List<KlineData> klines = new ArrayList<>();
-            BigDecimal basePrice = basePrices.getOrDefault(symbol.toUpperCase(Locale.ROOT), new BigDecimal("100.00"));
-            
-            Instant currentTime = endTime != null ? endTime : Instant.now();
-            if (endTime == null && startTime != null) {
-                currentTime = startTime;
-            }
-            Duration interval = parseTimeframe(timeframe);
-            
-            BigDecimal currentPrice = basePrice;
-            for (int i = 0; i < limit; i++) {
-                double changePercent = (ThreadLocalRandom.current().nextDouble() - 0.5) * 0.04;
-                BigDecimal change = currentPrice.multiply(BigDecimal.valueOf(changePercent));
-                BigDecimal close = currentPrice.add(change);
-                
-                BigDecimal high = close.multiply(
-                    BigDecimal.valueOf(1 + ThreadLocalRandom.current().nextDouble() * 0.01));
-                BigDecimal low = close.multiply(
-                    BigDecimal.valueOf(1 - ThreadLocalRandom.current().nextDouble() * 0.01));
-                BigDecimal open = currentPrice;
-                
-                long volume = ThreadLocalRandom.current().nextLong(100_000L, 10_000_000L);
-                
-                klines.add(KlineData.builder()
-                    .symbol(symbol.toUpperCase(Locale.ROOT))
-                    .market(MarketType.US_STOCK.getCode())
-                    .timestamp(currentTime)
-                    .open(open)
-                    .high(high)
-                    .low(low)
-                    .close(close)
-                    .volume(volume)
-                    .amount(close.multiply(BigDecimal.valueOf(volume)))
-                    .timeframe(timeframe)
-                    .build());
-                
-                currentPrice = close;
-                currentTime = currentTime.minus(interval);
-            }
-            
-            Collections.reverse(klines);
-            return klines;
-        }
-        
-        Optional<TickData> getRealTimeTick(String symbol) {
-            BigDecimal basePrice = basePrices.getOrDefault(symbol.toUpperCase(Locale.ROOT), new BigDecimal("100.00"));
-            
-            double changePercent = (ThreadLocalRandom.current().nextDouble() - 0.5) * 0.02;
-            BigDecimal price = basePrice.multiply(BigDecimal.valueOf(1 + changePercent));
-            BigDecimal change = price.subtract(basePrice);
-            BigDecimal changePercentValue = change.divide(basePrice, 4, RoundingMode.HALF_UP)
-                                                  .multiply(BigDecimal.valueOf(100));
-            
-            long volume = ThreadLocalRandom.current().nextLong(1_000_000L, 50_000_000L);
-            
-            TickData tickData = TickData.builder()
-                .symbol(symbol.toUpperCase(Locale.ROOT))
-                .market(MarketType.US_STOCK.getCode())
-                .timestamp(Instant.now())
-                .price(price)
-                .change(change)
-                .changePercent(changePercentValue)
-                .volume(volume)
-                .amount(price.multiply(BigDecimal.valueOf(volume)))
-                .bidPrice(price.multiply(BigDecimal.valueOf(0.999)))
-                .bidVolume(ThreadLocalRandom.current().nextLong(100L, 10_000L))
-                .askPrice(price.multiply(BigDecimal.valueOf(1.001)))
-                .askVolume(ThreadLocalRandom.current().nextLong(100L, 10_000L))
-                .dayHigh(price.multiply(BigDecimal.valueOf(1.02)))
-                .dayLow(price.multiply(BigDecimal.valueOf(0.98)))
-                .open(basePrice)
-                .prevClose(basePrice)
-                .build();
-            
-            return Optional.of(tickData);
-        }
-        
-        List<SymbolInfo> searchSymbols(String keyword, int limit) {
-            List<SymbolInfo> results = new ArrayList<>();
-            String upperKeyword = keyword.toUpperCase(Locale.ROOT);
-            
-            Map<String, String> usStocks = Map.of(
-                "AAPL", "Apple Inc.",
-                "MSFT", "Microsoft Corporation",
-                "GOOGL", "Alphabet Inc.",
-                "AMZN", "Amazon.com Inc.",
-                "TSLA", "Tesla Inc.",
-                "META", "Meta Platforms Inc.",
-                "NVDA", "NVIDIA Corporation",
-                "AMD", "Advanced Micro Devices",
-                "INTC", "Intel Corporation",
-                "NFLX", "Netflix Inc."
-            );
-            
-            usStocks.entrySet().stream()
-                .filter(e -> e.getKey().contains(upperKeyword) || 
-                            e.getValue().toUpperCase(Locale.ROOT).contains(upperKeyword))
-                .limit(limit)
-                .forEach(e -> results.add(new SymbolInfo(
-                    e.getKey(),
-                    e.getValue(),
-                    MarketType.US_STOCK.getCode(),
-                    "NASDAQ",
-                    "stock"
-                )));
-            
-            return results;
-        }
-        
-        Duration parseTimeframe(String timeframe) {
-            return switch (timeframe.toLowerCase(Locale.ROOT)) {
-                case "1m" -> Duration.ofMinutes(1);
-                case "5m" -> Duration.ofMinutes(5);
-                case "15m" -> Duration.ofMinutes(15);
-                case "30m" -> Duration.ofMinutes(30);
-                case "1h", "60m" -> Duration.ofHours(1);
-                case "1d", "daily" -> Duration.ofDays(1);
-                case "1w", "weekly" -> Duration.ofDays(7);
-                case "1mth", "monthly" -> Duration.ofDays(30);
-                default -> Duration.ofDays(1);
-            };
-        }
-    }
 }
