+++ koduck-backend/src/main/java/com/koduck/service/support/BacktestSignal.java
@@ -1,0 +1,13 @@
+package com.koduck.service.support;
+
+/**
+ * Trading signals generated during backtest simulation.
+ *
+ * @author GitHub Copilot
+ * @date 2026-03-31
+ */
+public enum BacktestSignal {
+    BUY,
+    SELL,
+    HOLD
+}
