+++ koduck-backend/src/main/java/com/koduck/service/support/MarketFallbackSupport.java
@@ -1,0 +1,223 @@
+package com.koduck.service.support;
+
+import com.koduck.common.constants.MarketConstants;
+import com.koduck.dto.market.PriceQuoteDto;
+import com.koduck.dto.market.StockIndustryDto;
+import com.koduck.dto.market.StockStatsDto;
+import com.koduck.dto.market.StockValuationDto;
+import com.koduck.dto.market.SymbolInfoDto;
+import com.koduck.entity.StockBasic;
+import com.koduck.market.MarketType;
+import com.koduck.market.provider.MarketDataProvider;
+import com.koduck.market.provider.ProviderFactory;
+import com.koduck.repository.StockBasicRepository;
+import com.koduck.service.KlineService;
+import com.koduck.service.market.AKShareDataProvider;
+import com.koduck.util.SymbolUtils;
+import java.math.BigDecimal;
+import java.time.Instant;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+/**
+ * Fallback helper for provider and kline based market query recovery.
+ *
+ * @author GitHub Copilot
+ * @date 2026-03-31
+ */
+@Component
+@Slf4j
+@RequiredArgsConstructor
+public class MarketFallbackSupport {
+
+    private static final String DEFAULT_MARKET = MarketConstants.DEFAULT_MARKET;
+    private static final String DAILY_TIMEFRAME = MarketConstants.DEFAULT_TIMEFRAME;
+    private static final String STOCK_TYPE = "STOCK";
+
+    private final StockBasicRepository stockBasicRepository;
+    private final KlineService klineService;
+    private final ProviderFactory providerFactory;
+    private final MarketServiceSupport marketServiceSupport;
+
+    public PriceQuoteDto fetchProviderPrice(String symbol) {
+        return getAShareProvider()
+            .flatMap(this::toAkShareProvider)
+            .map(provider -> provider.getPrice(symbol))
+            .orElse(null);
+    }
+
+    public List<SymbolInfoDto> searchSymbolsFromProvider(String keyword, int size) {
+        List<MarketDataProvider.SymbolInfo> providerResults = getAShareProvider()
+            .map(provider -> provider.searchSymbols(keyword, size))
+            .orElse(Collections.emptyList());
+        if (providerResults == null || providerResults.isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        Map<String, SymbolInfoDto> deduplicated = new LinkedHashMap<>();
+        for (MarketDataProvider.SymbolInfo symbolInfo : providerResults) {
+            if (symbolInfo == null || symbolInfo.symbol() == null || symbolInfo.symbol().isBlank()) {
+                continue;
+            }
+            String normalizedSymbol = SymbolUtils.normalize(symbolInfo.symbol());
+            String market = symbolInfo.market() == null || symbolInfo.market().isBlank()
+                ? DEFAULT_MARKET
+                : symbolInfo.market();
+            String name = symbolInfo.name() == null || symbolInfo.name().isBlank()
+                ? normalizedSymbol
+                : symbolInfo.name();
+
+            SymbolInfoDto dto = SymbolInfoDto.builder()
+                .symbol(normalizedSymbol)
+                .name(name)
+                .type(STOCK_TYPE)
+                .market(market)
+                .build();
+            deduplicated.putIfAbsent(market + ":" + normalizedSymbol, dto);
+        }
+
+        if (deduplicated.isEmpty()) {
+            return Collections.emptyList();
+        }
+        return deduplicated.values().stream().limit(size).toList();
+    }
+
+    public StockValuationDto fetchProviderValuation(String symbol) {
+        return getAShareProvider()
+            .flatMap(this::toAkShareProvider)
+            .map(provider -> provider.getStockValuation(symbol))
+            .orElse(null);
+    }
+
+    public StockIndustryDto fetchProviderIndustry(String symbol) {
+        return getAShareProvider()
+            .flatMap(this::toAkShareProvider)
+            .map(provider -> provider.getStockIndustry(symbol))
+            .orElse(null);
+    }
+
+    public PriceQuoteDto tryBuildQuoteFromLatestKline(String symbol) {
+        try {
+            return buildQuoteFromLatestKline(symbol);
+        } catch (Exception ex) {
+            log.warn("Failed to build quote from kline: symbol={}, error={}", symbol, ex.getMessage());
+            return null;
+        }
+    }
+
+    public StockStatsDto tryBuildStatsFromKline(String symbol, String market) {
+        try {
+            StockBasic basic = stockBasicRepository.findBySymbol(symbol).orElse(null);
+            if (basic == null) {
+                return null;
+            }
+
+            String actualMarket = market;
+            if (actualMarket == null || actualMarket.isBlank()) {
+                String basicMarket = basic.getMarket();
+                if (basicMarket != null && !basicMarket.isBlank()) {
+                    actualMarket = basicMarket;
+                } else {
+                    actualMarket = DEFAULT_MARKET;
+                }
+            }
+
+            List<com.koduck.dto.market.KlineDataDto> recent =
+                klineService.getKlineData(actualMarket, symbol, DAILY_TIMEFRAME, 2, null);
+            recent = marketServiceSupport.normalizeKlineData(recent);
+            if (recent.isEmpty()) {
+                return null;
+            }
+
+            com.koduck.dto.market.KlineDataDto latest = recent.get(recent.size() - 1);
+            BigDecimal prevClose = recent.size() >= 2 ? recent.get(recent.size() - 2).close() : null;
+            BigDecimal current = latest.close();
+            BigDecimal change = marketServiceSupport.calculateChange(current, prevClose);
+            BigDecimal changePercent = marketServiceSupport.calculateChangePercent(change, prevClose);
+
+            return StockStatsDto.builder()
+                .symbol(symbol)
+                .market(actualMarket)
+                .open(latest.open())
+                .high(latest.high())
+                .low(latest.low())
+                .current(current)
+                .prevClose(prevClose)
+                .change(change)
+                .changePercent(changePercent)
+                .volume(latest.volume())
+                .amount(latest.amount())
+                .timestamp(latest.timestamp() != null ? Instant.ofEpochSecond(latest.timestamp()) : null)
+                .build();
+        } catch (Exception ex) {
+            log.warn("Failed to build stats from kline: symbol={}, error={}", symbol, ex.getMessage());
+            return null;
+        }
+    }
+
+    private Optional<MarketDataProvider> getAShareProvider() {
+        return providerFactory.getPrimaryProvider(MarketType.A_SHARE);
+    }
+
+    private Optional<AKShareDataProvider> toAkShareProvider(MarketDataProvider provider) {
+        if (provider instanceof AKShareDataProvider akShareProvider) {
+            return Optional.of(akShareProvider);
+        }
+        log.warn("Primary provider for A_SHARE is not AKShareDataProvider: {}", provider.getProviderName());
+        return Optional.empty();
+    }
+
+    private PriceQuoteDto buildQuoteFromLatestKline(String symbol) {
+        StockBasic basic = stockBasicRepository.findBySymbol(symbol).orElse(null);
+        if (basic == null) {
+            return null;
+        }
+
+        String market = basic.getMarket() != null && !basic.getMarket().isBlank() ? basic.getMarket() : DEFAULT_MARKET;
+        PriceQuoteDto quote = buildQuoteFromKline(symbol, basic, market);
+        if (quote != null) {
+            return quote;
+        }
+        if (!DEFAULT_MARKET.equals(market)) {
+            return buildQuoteFromKline(symbol, basic, DEFAULT_MARKET);
+        }
+        return null;
+    }
+
+    private PriceQuoteDto buildQuoteFromKline(String symbol, StockBasic basic, String market) {
+        List<com.koduck.dto.market.KlineDataDto> recent =
+            klineService.getKlineData(market, symbol, DAILY_TIMEFRAME, 2, null);
+        recent = marketServiceSupport.normalizeKlineData(recent);
+        if (recent.isEmpty()) {
+            return null;
+        }
+
+        com.koduck.dto.market.KlineDataDto latest = recent.get(recent.size() - 1);
+        BigDecimal prevClose = recent.size() >= 2 ? recent.get(recent.size() - 2).close() : null;
+        BigDecimal price = latest.close();
+        BigDecimal change = marketServiceSupport.calculateChange(price, prevClose);
+        BigDecimal changePercent = marketServiceSupport.calculateChangePercent(change, prevClose);
+
+        return PriceQuoteDto.builder()
+            .symbol(symbol)
+            .name(basic.getName())
+            .type(STOCK_TYPE)
+            .price(price)
+            .open(latest.open())
+            .high(latest.high())
+            .low(latest.low())
+            .prevClose(prevClose)
+            .volume(latest.volume())
+            .amount(latest.amount())
+            .change(change)
+            .changePercent(changePercent)
+            .timestamp(latest.timestamp() != null ? Instant.ofEpochSecond(latest.timestamp()) : null)
+            .build();
+    }
+}
